+++ app/src/main/java/com/suvojeet/suvmusic/ui/screens/SeekbarStyleScreen.kt
@@ -268,7 +268,7 @@ private fun androidx.compose.ui.graphics.drawscope.DrawScope.drawWaveLinePreview
         var x = 0f
         while (x <= width) {
             val y = centerY + sin(x * 0.1f) * amplitude
-            lineTo(x, y.toFloat())
+            lineTo(x, y)
             x += 3f
         }
     }
@@ -280,7 +280,7 @@ private fun androidx.compose.ui.graphics.drawscope.DrawScope.drawWaveLinePreview
         var x = 0f
         while (x <= progressX) {
             val y = centerY + sin(x * 0.1f) * amplitude
-            lineTo(x, y.toFloat())
+            lineTo(x, y)
             x += 3f
         }
     }
